% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_two_way_crossed_ANOVA_package.R
\name{Psi_1}
\alias{Psi_1}
\title{To compute the test statistic value, critical value and the P-value of the test Psi1}
\usage{
Psi_1(a, b, nominalsize, cellsize, cellmean, cellvar)
}
\arguments{
\item{a}{number of levels of the row factor A}

\item{b}{number of levels of the column factor B}

\item{nominalsize}{the level of significance}

\item{cellsize}{matrix having a rows and b columns, and (i,j)-th element is the sample size in the (i,j)-th cell}

\item{cellmean}{matrix having a rows and b columns, and (i,j)-th element is the sample mean in the (i,j)-th cell}

\item{cellvar}{matrix having a rows and b columns, and (i,j)-th element is the sample variance in the (i,j)-th cell}
}
\value{
Numeric vector consisting of observed test statistic, critical and P-values of the test Psi1 and the decision of the test
}
\description{
To compute the test statistic value, critical value and the P-value of the test Psi1
}
\examples{
Psi_1(3,2,0.05,rbind(c(50,24),c(55,50),c(50,55)),rbind(c(69.8700,75.9833),c(73.2727,77.9140),c(77.4160,79.7090)),rbind(c(35.5678,10.6067),c(9.9028,12.9277),c(19.2650,15.0549)))
}
